Move SSR helper out of server API entrypoint
